chore/app: modify mdata API to take DirHandle instead of name and type_tag
