+++ src/app/ffi/mutable_data/helper.rs
@@ -21,8 +21,8 @@
 
 use app::{App, AppContext};
 use app::errors::AppError;
-use app::object_cache::{MDataPermissionsHandle, ObjectCache, SignKeyHandle};
-use core::{Client, FutureExt};
+use app::object_cache::{DirHandle, MDataPermissionsHandle, ObjectCache, SignKeyHandle};
+use core::{Client, Dir, FutureExt};
 use futures::Future;
 use routing::{PermissionSet, User};
 use std::collections::BTreeMap;
@@ -31,19 +31,16 @@ use std::os::raw::c_void;
 use util::ffi::OpaqueCtx;
 use util::ffi::callback::{Callback, CallbackArgs};
 
-// TODO: consider moving the send_* functions to ffi::helper, or even make them methods of
-// App.
-
-// Convenience wrapper around `App::send` which automatically handles the callback
-// boilerplate.
-// Use this if the lambda always returns future.
-pub unsafe fn send_async<C, F, U, E>(app: *const App,
-                                     user_data: *mut c_void,
-                                     cb: C,
-                                     f: F)
-                                     -> Result<(), AppError>
+// Helper to reduce boilerplate when sending asynchronous operations to the app
+// event loop.
+pub unsafe fn send_with_mdata_info<C, F, U, E>(app: *const App,
+                                               info_h: DirHandle,
+                                               user_data: *mut c_void,
+                                               cb: C,
+                                               f: F)
+                                               -> Result<(), AppError>
     where C: Callback + Copy + Send + 'static,
-          F: FnOnce(&Client, &AppContext) -> U + Send + 'static,
+          F: FnOnce(&Client, &AppContext, &Dir) -> U + Send + 'static,
           U: Future<Item = C::Args, Error = E> + 'static,
           E: Debug + 'static,
           AppError: From<E>
@@ -51,7 +48,8 @@ pub unsafe fn send_async<C, F, U, E>(app: *const App,
     let user_data = OpaqueCtx(user_data);
 
     (*app).send(move |client, context| {
-        f(client, context)
+        let info = try_cb!(context.object_cache().get_dir(info_h), user_data, cb);
+        f(client, context, &*info)
             .map(move |args| cb.call(user_data.0, 0, args))
             .map_err(AppError::from)
             .map_err(move |err| cb.call(user_data.0, ffi_error_code!(err), C::Args::default()))
