+++ src/core/client/account.rs
@@ -19,10 +19,9 @@
 // Please review the Licences for the specific language governing permissions
 // and limitations relating to use of the SAFE Network Software.
 
+use core::dir::Dir;
 use core::errors::CoreError;
-use core::utility::{symmetric_decrypt, symmetric_encrypt};
 use maidsafe_utilities::serialisation::{deserialise, serialise};
-use rand::{OsRng, Rng};
 use routing::{FullId, XOR_NAME_LEN, XorName};
 use rust_sodium::crypto::{box_, pwhash, secretbox, sign};
 use rust_sodium::crypto::hash::sha256;
@@ -113,78 +112,6 @@ impl Account {
     }
 }
 
-/// Information about a directory associated with an account.
-#[derive(Clone, Debug, PartialEq, RustcDecodable, RustcEncodable)]
-pub struct Dir {
-    /// Name of the data where the directory is stored.
-    pub name: XorName,
-    /// Type tag of the data where the directory is stored.
-    pub type_tag: u64,
-    /// Key to encrypt/decrypt the directory content.
-    /// and the nonce to be used for keys
-    pub enc_info: Option<(secretbox::Key, Option<secretbox::Nonce>)>,
-}
-
-impl Dir {
-    /// Generate random, private (encrypted) `Dir` with the given type tag.
-    pub fn random_private(type_tag: u64) -> Result<Self, CoreError> {
-        let mut rng = os_rng()?;
-        let enc_info = Some((secretbox::gen_key(), Some(secretbox::gen_nonce())));
-
-        Ok(Dir {
-            name: rng.gen(),
-            type_tag: type_tag,
-            enc_info: enc_info,
-        })
-    }
-    /// Generate a random, publicly accessible `Dir` with the given type tag
-    pub fn random_public(type_tag: u64) -> Result<Self, CoreError> {
-        let mut rng = os_rng()?;
-
-        Ok(Dir {
-            name: rng.gen(),
-            type_tag: type_tag,
-            enc_info: None,
-        })
-    }
-
-    /// encrypt the the key for the mdata entry of this dir accordingly
-    pub fn enc_entry_key(&self, plain_text: &[u8]) -> Result<Vec<u8>, CoreError> {
-        if let Some((ref key, seed)) = self.enc_info {
-            let nonce = match seed {
-                Some(secretbox::Nonce(ref dir_nonce)) => {
-                    let mut pt = plain_text.to_vec();
-                    pt.extend_from_slice(&dir_nonce[..]);
-                    unwrap!(secretbox::Nonce::from_slice(
-                        &sha256::hash(&pt)[..secretbox::NONCEBYTES]))
-                }
-                None => secretbox::gen_nonce(),
-            };
-            Ok(serialise(&(nonce, secretbox::seal(plain_text, &nonce, key)))?)
-        } else {
-            Ok(plain_text.to_vec())
-        }
-    }
-
-    /// encrypt the value for this mdata entry accordingly
-    pub fn enc_entry_value(&self, plain_text: &[u8]) -> Result<Vec<u8>, CoreError> {
-        if let Some((ref key, _)) = self.enc_info {
-            symmetric_encrypt(plain_text, key)
-        } else {
-            Ok(plain_text.to_vec())
-        }
-    }
-
-    /// decrypt key or value of this mdata entry
-    pub fn decrypt(&self, cipher: &[u8]) -> Result<Vec<u8>, CoreError> {
-        if let Some((ref key, _)) = self.enc_info {
-            symmetric_decrypt(cipher, key)
-        } else {
-            Ok(cipher.to_vec())
-        }
-    }
-}
-
 /// Client signing and encryption keypairs
 #[derive(Clone, Debug, PartialEq, RustcDecodable, RustcEncodable)]
 pub struct ClientKeys {
@@ -225,15 +152,10 @@ impl Into<FullId> for ClientKeys {
     }
 }
 
-fn os_rng() -> Result<OsRng, CoreError> {
-    OsRng::new().map_err(|_| CoreError::RandomDataGenerationFailure)
-}
-
 #[cfg(test)]
 mod tests {
+    use core::dir::Dir;
     use maidsafe_utilities::serialisation::{deserialise, serialise};
-    use rand;
-    use rust_sodium::crypto::secretbox;
     use std::u32;
     use super::*;
 
@@ -310,43 +232,6 @@ mod tests {
 
         let decrypted = unwrap!(Account::decrypt(&encrypted, password, pin));
         assert_eq!(account, decrypted);
-    }
-
-    #[test]
-    fn random_dir_encrypts() {
-        let dir = unwrap!(Dir::random_private(0));
-        let key = Vec::from("str of key");
-        let val = Vec::from("other is value");
-        let enc_key = unwrap!(dir.enc_entry_key(&key));
-        let enc_val = unwrap!(dir.enc_entry_value(&val));
-        assert_ne!(enc_key, key);
-        assert_ne!(enc_val, val);
-        assert_eq!(unwrap!(dir.decrypt(&enc_key)), key);
-        assert_eq!(unwrap!(dir.decrypt(&enc_val)), val);
-    }
-
-    #[test]
-    fn public_dir_doesnt_encrypt() {
-        let dir = unwrap!(Dir::random_public(0));
-        let key = Vec::from("str of key");
-        let val = Vec::from("other is value");
-        assert_eq!(unwrap!(dir.enc_entry_key(&key)), key);
-        assert_eq!(unwrap!(dir.enc_entry_value(&val)), val);
-        assert_eq!(unwrap!(dir.decrypt(&val)), val);
-    }
-
-    #[test]
-    fn no_nonce_means_random_nonce() {
-        let dir = Dir {
-            name: rand::random(),
-            type_tag: 0,
-            enc_info: Some((secretbox::gen_key(), None)),
-        };
-        let key = Vec::from("str of key");
-        let enc_key = unwrap!(dir.enc_entry_key(&key));
-        assert_ne!(enc_key, key);
-        // encrypted is different on every run
-        assert_ne!(unwrap!(dir.enc_entry_key(&key)), key);
     }
 
     fn create_account() -> Account {
