+++ src/util/ffi/macros.rs
@@ -44,6 +44,8 @@ macro_rules! try_cb {
         match $result {
             Ok(value) => value,
             Err(err) => {
+                #[cfg_attr(feature = "clippy", allow(useless_attribute))]
+                #[allow(unused)]
                 use $crate::util::ffi::callback::{Callback, CallbackArgs};
                 $cb.call($user_data.into(), ffi_error_code!(err), CallbackArgs::default());
                 return None;
